+++ gslib/commands/du.py
@@ -27,9 +27,9 @@ from gslib.exception import CommandException
 from gslib.ls_helper import LsHelper
 from gslib.storage_url import ContainsWildcard
 from gslib.storage_url import StorageUrlFromString
-from gslib.util import MakeHumanReadable
 from gslib.util import NO_MAX
 from gslib.util import UTF8
+from gslib.utils.unit_util import MakeHumanReadable
 
 _SYNOPSIS = """
   gsutil du url...
