+++ gslib/commands/hash.py
@@ -38,7 +38,7 @@ from gslib.storage_url import StorageUrlFromString
 from gslib.thread_message import FileMessage
 from gslib.thread_message import FinalMessage
 from gslib.util import NO_MAX
-from gslib.util import UsingCrcmodExtension
+from gslib.utils.boto_util import UsingCrcmodExtension
 
 _SYNOPSIS = """
   gsutil hash [-c] [-h] [-m] filename...
