+++ gslib/commands/rsync.py
@@ -78,7 +78,6 @@ from gslib.storage_url import StorageUrlFromString
 from gslib.third_party.storage_apitools import storage_v1_messages as apitools_messages
 from gslib.translation_helper import CopyCustomMetadata
 from gslib.translation_helper import GenerationFromUrlAndString
-from gslib.util import CalculateThroughput
 from gslib.util import CreateCustomMetadata
 from gslib.util import CreateLock
 from gslib.util import GetCloudApiInstance
@@ -86,12 +85,13 @@ from gslib.util import GetValueFromObjectCustomMetadata
 from gslib.util import IS_WINDOWS
 from gslib.util import IsCloudSubdirPlaceholder
 from gslib.util import ObjectIsGzipEncoded
-from gslib.util import SECONDS_PER_DAY
-from gslib.util import TEN_MIB
-from gslib.util import UsingCrcmodExtension
 from gslib.util import UTF8
+from gslib.utils.boto_util import UsingCrcmodExtension
 from gslib.utils.rsync_util import DiffAction
 from gslib.utils.rsync_util import RsyncDiffToApply
+from gslib.utils.unit_util import CalculateThroughput
+from gslib.utils.unit_util import SECONDS_PER_DAY
+from gslib.utils.unit_util import TEN_MIB
 from gslib.wildcard_iterator import CreateWildcardIterator
