+++ gslib/gcs_json_api.py
@@ -81,16 +81,16 @@ from gslib.translation_helper import DEFAULT_CONTENT_TYPE
 from gslib.translation_helper import PRIVATE_DEFAULT_OBJ_ACL
 from gslib.translation_helper import REMOVE_CORS_CONFIG
 from gslib.util import AddAcceptEncodingGzipIfNeeded
-from gslib.util import GetCertsFile
-from gslib.util import GetCredentialStoreFilename
-from gslib.util import GetJsonResumableChunkSize
-from gslib.util import GetMaxRetryDelay
-from gslib.util import GetNewHttp
-from gslib.util import GetNumRetries
 from gslib.util import GetPrintableExceptionString
-from gslib.util import JsonResumableChunkSizeDefined
 from gslib.util import LogAndHandleRetries
 from gslib.util import NUM_OBJECTS_PER_LIST_PAGE
+from gslib.utils.boto_util import GetCertsFile
+from gslib.utils.boto_util import GetCredentialStoreFilename
+from gslib.utils.boto_util import GetJsonResumableChunkSize
+from gslib.utils.boto_util import GetMaxRetryDelay
+from gslib.utils.boto_util import GetNewHttp
+from gslib.utils.boto_util import GetNumRetries
+from gslib.utils.boto_util import JsonResumableChunkSizeDefined
 
 import httplib2
 import oauth2client
